add journey map
I tried to vary risk and reward, and tried a few symmetrical maps before setting on one asymmetrical. There's a couple mines kinda near the wells, and five more that are far away. There's some obstacles to create alternate paths, and I otherwise tried to avoid making narrow chokepoints or low traffic areas. The two main wells are are exposed in all directions to make them hard to monopolize, and the third is just remote, furthest from the mines.